requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
cffi==1.17.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
fonttools==4.58.5
kiwisolver==1.4.8
matplotlib==3.10.3
mplcursors==0.6
narwhals==1.45.0
numpy==2.3.1
packaging==25.0
paramiko==3.5.1
pefile==2023.2.7
pigpio==1.78
pillow==11.3.0
plotly==6.2.0
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
scp==0.15.0
setuptools==80.9.0
six==1.17.0
